requirements.txt
appnope==0.1.3
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.3.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cssselect==1.1.0
debugpy==1.6.0
decorator==5.1.1
Django==4.0.4
django-colorfield==0.6.3
django-extensions==3.1.5
django-json-widget==1.1.1
entrypoints==0.4
executing==0.8.3
future==0.18.2
idna==3.3
ipykernel==6.13.0
ipython==8.2.0
jedi==0.18.1
jupyter-client==7.2.2
jupyter-core==4.10.0
lxml==4.8.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.5
numpy==1.22.3
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.8
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2022.1
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sportsipy==0.6.0
sqlparse==0.4.2
stack-data==0.2.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
whitenoise==6.0.0
